// generated/src/apis/PlaygroundApi.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SegmentationAPI
 * API for asynchronous image and video segmentation and dashboard management.
 *
 * The version of the OpenAPI document: 2026-08-03
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import * as runtime from '../runtime.js';
import {
    type ApiError,
    ApiErrorFromJSON,
    ApiErrorToJSON,
} from '../models/ApiError.js';
import {
    type AsyncAcceptedResponse,
    AsyncAcceptedResponseFromJSON,
    AsyncAcceptedResponseToJSON,
} from '../models/AsyncAcceptedResponse.js';
import {
    type PlaygroundJobCreateRequest,
    PlaygroundJobCreateRequestFromJSON,
    PlaygroundJobCreateRequestToJSON,
} from '../models/PlaygroundJobCreateRequest.js';

export interface PlaygroundApiCreatePlaygroundJobRequest {
    playgroundJobCreateRequest: PlaygroundJobCreateRequest;
}

/**
 * 
 */
export class PlaygroundApi extends runtime.BaseAPI {

    /**
     * Creates request options for createPlaygroundJob without sending the request
     */
    async createPlaygroundJobRequestOpts(requestParameters: PlaygroundApiCreatePlaygroundJobRequest): Promise<runtime.RequestOpts> {
        if (requestParameters['playgroundJobCreateRequest'] == null) {
            throw new runtime.RequiredError(
                'playgroundJobCreateRequest',
                'Required parameter "playgroundJobCreateRequest" was null or undefined when calling createPlaygroundJob().'
            );
        }

        const queryParameters: any = {};

        const headerParameters: runtime.HTTPHeaders = {};

        headerParameters['Content-Type'] = 'application/json';

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["x-api-key"] = await this.configuration.apiKey("x-api-key"); // SharedAuthorizer authentication
        }

        if (this.configuration && this.configuration.accessToken) {
            const token = this.configuration.accessToken;
            const tokenString = await token("DashboardJwt", []);

            if (tokenString) {
                headerParameters["Authorization"] = `Bearer ${tokenString}`;
            }
        }

        let urlPath = `/v1/playground/jobs`;

        return {
            path: urlPath,
            method: 'POST',
            headers: headerParameters,
            query: queryParameters,
            body: PlaygroundJobCreateRequestToJSON(requestParameters['playgroundJobCreateRequest']),
        };
    }

    /**
     * Create an asynchronous single-image playground segmentation job
     */
    async createPlaygroundJobRaw(requestParameters: PlaygroundApiCreatePlaygroundJobRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<AsyncAcceptedResponse>> {
        const requestOptions = await this.createPlaygroundJobRequestOpts(requestParameters);
        const response = await this.request(requestOptions, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => AsyncAcceptedResponseFromJSON(jsonValue));
    }

    /**
     * Create an asynchronous single-image playground segmentation job
     */
    async createPlaygroundJob(requestParameters: PlaygroundApiCreatePlaygroundJobRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<AsyncAcceptedResponse> {
        const response = await this.createPlaygroundJobRaw(requestParameters, initOverrides);
        return await response.value();
    }

}